fix for tasks commit

+++ problems/arrays/array_smoothing_c/solution.cpp
@@ -3,14 +3,14 @@
 
 int* read_array(int size) {
     int* arr = (int*)malloc(size * sizeof(int));
-    for (size_t i = 0; i < size; ++i) {
+    for (int i = 0; i < size; ++i) {
         scanf("%d", &arr[i]);
     }
     return arr;
 }
 
 void print_array(int* array, int size) {
-	for (size_t i = 0; i < size; ++i) {
+	for (int i = 0; i < size; ++i) {
 		printf("%d ", array[i]);
 	}
 }
@@ -25,7 +25,7 @@ void make_smoothing(int** arr, int size) {
 
 	smoothing[0] = get_average(array[size-1], array[0], array[1]);
 	smoothing[size - 1] = get_average(array[size-2], array[size-1], array[0]);
-	for (size_t i = 1; i < size-1; ++i) {
+	for (int i = 1; i < size-1; ++i) {
 		smoothing[i] = get_average(array[i-1], array[i], array[i+1]);
 	}
 	free(array);
